Move `PlayerListEntry` to packets.ts to resolve dependency cycle
